Fixed more warnings with GCC and Clang on Linux, and verified no new Visual Studio warnings worth fixing showed up from the changes (progress on #24).
One of the warnings, around token pasting a comma to __VA_ARGS__ in a macro, won't be fixed, because that non-standard extension works on all the supported platforms/compilers. Though we could change how error handling works so __VA_ARGS__ would no longer be used.

At this point, with the code not yet fully refactored into C++, much of the code simply can't be made to not show some of the warnings worth fixing with simple changes (like all the unused function parameters). That means compiling with -Werror isn't possible, for now. So I don't consider this commit fully resolving #24.

+++ src/SGUIL/SGUIL_GuiTextField.cc
@@ -277,13 +277,13 @@ void GuiTextField::mouseClicked(int x, int y, Uint8 button)
         updatePositionalValues = false;
     }
 
-    int pos = getPositionUnderMouse(x, y);
+    unsigned pos = getPositionUnderMouse(x, y);
     if(pos == TEXT_POSITION_NONE)
     {
         return;
     }
 
-    if(SDL_GetModState() & KMOD_SHIFT && unsigned(pos) != cursor)
+    if(SDL_GetModState() & KMOD_SHIFT && pos != cursor)
     {
         selectionStart = cursor;
         selectionEnd = pos;
@@ -308,7 +308,7 @@ void GuiTextField::mouseDragged(int x, int y, Uint8 button)
         updatePositionalValues = false;
     }
 
-    int pos = getPositionUnderMouse(x, y);
+    unsigned pos = getPositionUnderMouse(x, y);
     if(pos == TEXT_POSITION_NONE)
     {
         return;
